requirements.txt
click==8
azure-storage-blob==12.11.0
streamlit==1.8.1